+++ runtime/include/dart_api_dl.c
@@ -10,25 +10,27 @@
 
 #include <string.h>
 
-#define DART_API_DL_DEFINITIONS(name)                                          \
-  using name##Type = decltype(&name);                                          \
-  name##Type name##_DL = nullptr;
+#define DART_API_DL_DEFINITIONS(name, R, A)                                    \
+  typedef R(*name##_Type) A;                                                   \
+  name##_Type name##_DL = NULL;
+
 DART_API_ALL_DL_SYMBOLS(DART_API_DL_DEFINITIONS)
+
 #undef DART_API_DL_DEFINITIONS
 
 typedef void (*DartApiEntry_function)();
 
 DartApiEntry_function FindFunctionPointer(const DartApiEntry* entries,
                                           const char* name) {
-  while (entries->name != nullptr) {
+  while (entries->name != NULL) {
     if (strcmp(entries->name, name) == 0) return entries->function;
     entries++;
   }
-  return nullptr;
+  return NULL;
 }
 
 intptr_t Dart_InitializeApiDL(void* data) {
-  DartApi* dart_api_data = reinterpret_cast<DartApi*>(data);
+  DartApi* dart_api_data = (DartApi*)data;
 
   if (dart_api_data->major != DART_API_DL_MAJOR_VERSION) {
     // If the DartVM we're running on does not have the same version as this
@@ -49,9 +51,9 @@ intptr_t Dart_InitializeApiDL(void* data) {
 
   const DartApiEntry* dart_api_function_pointers = dart_api_data->functions;
 
-#define DART_API_DL_INIT(name)                                                 \
-  name##_DL = reinterpret_cast<name##Type>(                                    \
-      FindFunctionPointer(dart_api_function_pointers, #name));
+#define DART_API_DL_INIT(name, R, A)                                           \
+  name##_DL =                                                                  \
+      (name##_Type)(FindFunctionPointer(dart_api_function_pointers, #name));
   DART_API_ALL_DL_SYMBOLS(DART_API_DL_INIT)
 #undef DART_API_DL_INIT